project/requirements.txt
numpy
pandas
nltk
transformers[pytorch]
pytorch
pytorch-lightning
detoxify
vision-models-playground
scikit-learn
seaborn
ipywidgets
notebook
googletrans==4.0.0rc1
textaugment
nlpaug
sentencepiece
sacremoses
dotenv
tweepy